LayoutTests/fast/dom/Window/new-window-opener-expected.txt

This tests that the different arguments given to a new window are honored.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Scrollbars
PASS newWin.scrollbars.visible is false
FAIL newWin.scrollbars.visible should be true. Was false.
PASS newWin.scrollbars.visible is false
FAIL newWin.scrollbars.visible should be true. Was false.

Toolbars
PASS newWin.toolbar.visible is false
FAIL newWin.toolbar.visible should be true. Was false.
PASS newWin.toolbar.visible is false
FAIL newWin.toolbar.visible should be true. Was false.

Statusbar
PASS newWin.statusbar.visible is false
FAIL newWin.statusbar.visible should be true. Was false.
PASS newWin.statusbar.visible is false
FAIL newWin.statusbar.visible should be true. Was false.

Menubar
FAIL newWin.menubar.visible should be false. Was true.
PASS newWin.menubar.visible is true
FAIL newWin.menubar.visible should be false. Was true.
PASS newWin.menubar.visible is true

Locationbar
PASS newWin.locationbar.visible is false
FAIL newWin.locationbar.visible should be true. Was false.
PASS newWin.locationbar.visible is false
FAIL newWin.locationbar.visible should be true. Was false.

All together now!
FAIL newWin.scrollbars.visible should be true. Was false.
FAIL newWin.toolbar.visible should be true. Was false.
FAIL newWin.statusbar.visible should be true. Was false.
PASS newWin.menubar.visible is true
FAIL newWin.locationbar.visible should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
